14350 Extract ItemListState to new file
