+++ WooCommerce/Classes/POS/ViewModels/ItemListViewModelProtocol.swift
@@ -4,13 +4,13 @@ import protocol Yosemite.POSItem
 
 protocol ItemListViewModelProtocol: ObservableObject {
     var items: [POSItem] { get }
-    var state: ItemListViewModel.ItemListState { get }
+    var state: ItemListState { get }
     var isHeaderBannerDismissed: Bool { get }
     var shouldShowHeaderBanner: Bool { get }
 
     var selectedItemPublisher: AnyPublisher<POSItem, Never> { get }
     var itemsPublisher: Published<[POSItem]>.Publisher { get }
-    var statePublisher: Published<ItemListViewModel.ItemListState>.Publisher { get }
+    var statePublisher: Published<ItemListState>.Publisher { get }
 
     func select(_ item: POSItem)
     func loadInitialItems() async
